fix: No sigterm screaming
