+++ src/braket/simulator_v2/__init__.py
@@ -1,7 +1,6 @@
 from braket.simulator_v2.density_matrix_simulator_v2 import (  # noqa: F401
     DensityMatrixSimulatorV2,
 )
-from braket.simulator_v2.julia_import import setup_julia  # noqa: F401
 from braket.simulator_v2.state_vector_simulator_v2 import (  # noqa: F401
     StateVectorSimulatorV2,
 )
