+++ tox.ini
@@ -10,9 +10,9 @@ allowlist_externals =
     pytest
 setenv =
     PYTHON_JULIACALL_HANDLE_SIGNALS=yes
-    PYTHON_JULIACALL_OPTIMIZE=0
+    PYTHON_JULIAPKG_OFFLINE=yes
 commands =
-    python -c 'import juliacall' # force install of all deps
+    python -c 'import juliacall' # force install of all deps, trigger precompilation
     pytest {posargs} --cov-report term-missing --cov-report html --cov-report xml --cov=braket
 extras = test
